ux: gcc-10 fix

+++ src/kern/ux/irq_chip_ux_base.cpp
@@ -150,7 +150,7 @@ Irq_chip_ux_base::setup_irq_prov(unsigned irq, const char *const path,
 
   switch (pids[irq] = fork())
     {
-      case -1:
+      case (unsigned int)-1:
         return false;
 
       case 0:
